[추가] 이미지 팝업시 이미지 줌
- text, image NIB을 분리했어요

- 아직 말풍선에서 이미지크기를 조정하는 방법은 못찾았고
  지금 4:3 비율로 맞춰놓은 상태에요
