+++ Project_SOS/Project_SOS/BY_DetailTableViewCell.swift
@@ -8,9 +8,6 @@
 
 import UIKit
 
-protocol bubbleImageCellDelegate {
-    func presentPopup()
-}
 
 class BY_DetailTableViewCell: UITableViewCell {
 
@@ -19,19 +16,10 @@ class BY_DetailTableViewCell: UITableViewCell {
     //MARK:-        Properties                 //
     /*******************************************/
     
-    var delegate:bubbleImageCellDelegate?
-    
-    @IBOutlet weak var clickedImageOutlet: UIButton!
+
     @IBOutlet weak var bgView: UIView!
     @IBOutlet weak var characterIconImage: UIImageView!
     @IBOutlet weak var explainBubbleText: UILabel!
-    @IBOutlet weak var explainBubbleImage: UIImageView!
-
-    
-    @IBAction func clickedImageBtn(_ sender: UIButton) {
-        print("이미지셀이 눌렸다능")
-        self.delegate?.presentPopup()
-    }
 
     
     //Explain ImageView Constraints
@@ -50,8 +38,6 @@ class BY_DetailTableViewCell: UITableViewCell {
     override func awakeFromNib() {
         super.awakeFromNib()
         self.bgView.layer.cornerRadius = 10
-        self.explainBubbleImage.layer.cornerRadius = 10
-        self.explainBubbleImage.clipsToBounds = true
     }
     
     override func layoutSubviews() {
